be/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
asgiref=3.8.1=pyhd8ed1ab_0
blas=1.0=openblas
bottleneck=1.3.7=py312ha86b861_0
brotli-python=1.0.9=py312h313beb8_7
bzip2=1.0.8=h80987f9_5
ca-certificates=2024.3.11=hca03da5_0
certifi=2024.2.2=py312hca03da5_0
cffi=1.16.0=py312h80987f9_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cryptography=42.0.5=py312hd4332d6_0
defusedxml=0.7.1=pyhd3eb1b0_0
django=4.1=py312hca03da5_0
django-cors-headers=4.3.1=pyhd8ed1ab_0
django-pandas=0.6.6=pyhd8ed1ab_0
djangorestframework=3.15.1=pyhd8ed1ab_0
expat=2.5.0=h313beb8_0
httpie=3.2.1=pyhe4f9e05_1
idna=3.4=py312hca03da5_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
llvm-openmp=14.0.6=hc6e5704_0
markdown-it-py=2.2.0=py312hca03da5_1
mdurl=0.1.0=py312hca03da5_0
multidict=6.0.4=py312h80987f9_0
ncurses=6.4=h313beb8_0
numexpr=2.8.7=py312h0f3ea24_0
numpy=1.26.4=py312h7f4fdc5_0
numpy-base=1.26.4=py312he047099_0
openssl=3.0.13=h1a28f6b_0
pandas=2.2.1=py312hd77ebd4_0
pip=23.3.1=py312hca03da5_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pyhd8ed1ab_0
pyopenssl=24.0.0=py312hca03da5_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.2=h99e199e_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py312hca03da5_0
readline=8.2=h1a28f6b_0
requests=2.31.0=py312hca03da5_1
requests-toolbelt=1.0.0=py312hca03da5_0
rich=13.3.5=py312hca03da5_1
setuptools=68.2.2=py312hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h80987f9_0
sqlparse=0.4.4=py312hca03da5_0
tk=8.6.12=hb8d0fd4_0
typing_extensions=4.9.0=py312hca03da5_1
tzdata=2024a=h04d1e81_0
urllib3=1.26.18=py312hca03da5_0
wheel=0.41.2=py312hca03da5_0
xz=5.4.6=h80987f9_0
zlib=1.2.13=h5a0b063_0
